refactor: reorganize SafeTeleport to land/internal namespace
