+++ src/pland/PLand.cpp
@@ -16,12 +16,12 @@
 #include "drawer/DrawHandleManager.h"
 #include "events/domain/ConfigReloadEvent.h"
 #include "land/internal/LandScheduler.h"
+#include "land/internal/SafeTeleport.h"
 #include "pland/adapter/telemetry/Telemetry.h"
 #include "pland/command/Command.h"
 #include "pland/economy/EconomySystem.h"
 #include "pland/hooks/EventListener.h"
 #include "pland/infra/Config.h"
-#include "pland/infra/SafeTeleport.h"
 #include "pland/land/repo/LandRegistry.h"
 #include "pland/selector/SelectorManager.h"
 #include "pland/service/ServiceLocator.h"
@@ -38,9 +38,9 @@ struct PLand::Impl {
     ll::mod::NativeMod&                             mSelf;
     std::unique_ptr<ll::thread::ThreadPoolExecutor> mThreadPoolExecutor{nullptr};
     std::unique_ptr<LandRegistry>                   mLandRegistry{nullptr};
-    std::unique_ptr<LandScheduler>                  mLandScheduler{nullptr};
+    std::unique_ptr<internal::LandScheduler>        mLandScheduler{nullptr};
     std::unique_ptr<EventListener>                  mEventListener{nullptr};
-    std::unique_ptr<SafeTeleport>                   mSafeTeleport{nullptr};
+    std::unique_ptr<internal::SafeTeleport>         mSafeTeleport{nullptr};
     std::unique_ptr<SelectorManager>                mSelectorManager{nullptr};
     std::unique_ptr<DrawHandleManager>              mDrawHandleManager{nullptr};
     std::unique_ptr<adapter::Telemetry>             mTelemetry{nullptr};
@@ -86,9 +86,9 @@ bool PLand::load() {
 
 bool PLand::enable() {
     LandCommand::setup();
-    mImpl->mLandScheduler     = std::make_unique<LandScheduler>();
+    mImpl->mLandScheduler     = std::make_unique<internal::LandScheduler>();
     mImpl->mEventListener     = std::make_unique<EventListener>();
-    mImpl->mSafeTeleport      = std::make_unique<SafeTeleport>();
+    mImpl->mSafeTeleport      = std::make_unique<internal::SafeTeleport>();
     mImpl->mSelectorManager   = std::make_unique<SelectorManager>();
     mImpl->mDrawHandleManager = std::make_unique<DrawHandleManager>();
     mImpl->mTelemetry         = std::make_unique<adapter::Telemetry>();
@@ -164,11 +164,11 @@ PLand& PLand::getInstance() {
 
 PLand::PLand() : mImpl(std::make_unique<Impl>()) {}
 
-ll::mod::NativeMod& PLand::getSelf() const { return mImpl->mSelf; }
-SafeTeleport*       PLand::getSafeTeleport() const { return mImpl->mSafeTeleport.get(); }
-SelectorManager*    PLand::getSelectorManager() const { return mImpl->mSelectorManager.get(); }
-LandRegistry&       PLand::getLandRegistry() const { return *mImpl->mLandRegistry; }
-DrawHandleManager*  PLand::getDrawHandleManager() const { return mImpl->mDrawHandleManager.get(); }
+ll::mod::NativeMod&     PLand::getSelf() const { return mImpl->mSelf; }
+SelectorManager*        PLand::getSelectorManager() const { return mImpl->mSelectorManager.get(); }
+LandRegistry&           PLand::getLandRegistry() const { return *mImpl->mLandRegistry; }
+DrawHandleManager*      PLand::getDrawHandleManager() const { return mImpl->mDrawHandleManager.get(); }
+internal::SafeTeleport& PLand::getSafeTeleport() const { return *mImpl->mSafeTeleport; }
 
 ll::thread::ThreadPoolExecutor* PLand::getThreadPool() const { return mImpl->mThreadPoolExecutor.get(); }
 service::ServiceLocator&        PLand::getServiceLocator() const { return *mImpl->mServiceLocator; }
