+++ src/pland/PLand.h
@@ -28,6 +28,11 @@ class DevToolApp;
 
 namespace land {
 
+namespace internal {
+class SafeTeleport;
+}
+
+
 class PLand {
     PLand();
 
@@ -42,7 +47,6 @@ public: /* private */
 public: /* public */
     LDAPI static PLand& getInstance();
 
-    LDNDAPI class SafeTeleport*      getSafeTeleport() const;
     LDNDAPI class SelectorManager*   getSelectorManager() const;
     LDNDAPI class LandRegistry&      getLandRegistry() const;
     LDNDAPI class DrawHandleManager* getDrawHandleManager() const;
@@ -50,6 +54,8 @@ public: /* public */
     LDNDAPI ll::thread::ThreadPoolExecutor* getThreadPool() const;
 
     LDNDAPI service::ServiceLocator& getServiceLocator() const;
+
+    internal::SafeTeleport& getSafeTeleport() const;
 
 #ifdef LD_DEVTOOL
     [[nodiscard]] devtool::DevToolApp* getDevToolApp() const;
