+++ src/pland/land/internal/SafeTeleport.h
@@ -1,0 +1,21 @@
+#pragma once
+#include "pland/Global.h"
+
+class BlockPos;
+class Player;
+
+namespace land::internal {
+
+class SafeTeleport {
+public:
+    explicit SafeTeleport();
+    ~SafeTeleport();
+
+    void launchTask(Player& player, BlockPos const& targetPos, DimensionType dimension);
+
+private:
+    struct Impl;
+    std::unique_ptr<Impl> impl;
+};
+
+} // namespace land::internal
